backend/requirements.txt
pytest
pytest-asyncio
httpx
fastapi
uvicorn
python-docx
playwright
playwright-stealth
duckduckgo-search
python-multipart
thefuzz
sqlalchemy
python-jose[cryptography]
bcrypt
